Crate swc_ecma_transforms_testing

Modules

babel_like

Macros

compare_stdout

Test transformation by invoking it using node. The code must print something to stdout.

test

test_exec

Test transformation.

test_inline

Structs

FixtureTestConfig

Config for test_fixture. See test_fixture for documentation.

HygieneTester

Converts foo#1 to foo__1 so it can be verified by the test.

HygieneVisualizer

Tester

Functions

compare_stdout

Execute node for input and ensure that it prints same output after transformation.

exec_tr

Execute jest after transpiling input using tr.

parse_options

Just like babel, walk up the directory tree and find a file named options.json.

test_fixture

You can do UPDATE=1 cargo test to update fixtures.

test_transform